requirements.txt
asgiref==3.5.2
autopep8==2.0.0
defusedxml==0.7.1
Django==4.1.3
fonttools==4.38.0
fpdf2==2.6.0
Pillow==9.3.0
pycodestyle==2.10.0
sqlparse==0.4.3
tomli==2.0.1
